Move forge's mixin package to the same format as quilt
