+++ forge/src/main/java/com/rpmtw/rpmtw_platform_mod/mixins/forge/MixinBookRegistry.java
@@ -1,7 +1,7 @@
 /*
 Credit: https://github.com/kappa-maintainer/PRP
  */
-package com.rpmtw.rpmtw_platform_mod.forge.mixins;
+package com.rpmtw.rpmtw_platform_mod.mixins.forge;
 
 import net.minecraft.resources.ResourceLocation;
 import org.spongepowered.asm.mixin.Final;
