+++ forge/src/main/resources/rpmtw_platform_mod.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "com.rpmtw.rpmtw_platform_mod.forge.mixins",
+  "package": "com.rpmtw.rpmtw_platform_mod.mixins.forge",
   "compatibilityLevel": "JAVA_17",
   "injectors": {
     "defaultRequire": 1
